refactor: logica favorite state fora da home page
