+++ app_movies/lib/pages/Home/home_page.dart
@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:carousel_slider/carousel_slider.dart';
-import 'package:iconsax/iconsax.dart';
 import 'package:app_movies/core/app_colors.dart';
 import 'package:app_movies/pages/Home/home_details_screen.dart';
 import 'package:app_movies/api/movie_api.dart';
 import 'package:app_movies/models/movie.dart';
 import 'package:app_movies/pages/Login/login_page.dart';
-import 'package:provider/provider.dart';
-import 'package:app_movies/pages/favorite/favorites_provider.dart';
+import 'package:app_movies/pages/favorite/favorite_icon_state.dart';
 
 class MyHomePage extends StatefulWidget {
   const MyHomePage({super.key});
@@ -323,34 +321,5 @@ class _MyHomePageState extends State<MyHomePage>
             ),
           );
         });
-  }
-}
-
-class FavoriteIconButton extends StatefulWidget {
-  final Movie movie;
-
-  const FavoriteIconButton({super.key, required this.movie});
-
-  @override
-  State<FavoriteIconButton> createState() => _FavoriteIconButtonState();
-}
-
-class _FavoriteIconButtonState extends State<FavoriteIconButton> {
-  @override
-  Widget build(BuildContext context) {
-    return IconButton(
-      onPressed: () {
-        final favoritesProvider =
-            Provider.of<FavoritesProvider>(context, listen: false);
-        favoritesProvider.toggleFavorite(widget.movie);
-      },
-      icon: Icon(
-        context.watch<FavoritesProvider>().isFavorite(widget.movie.id)
-            ? Iconsax.heart5
-            : Iconsax.heart,
-        size: 24,
-        color: Colors.red,
-      ),
-    );
   }
 }
